day_8/requirements.txt
aiohttp==3.13.3
python-dotenv==1.0.1
tokenizers>=0.19.1